Remove all blocking code.
