+++ .github/workflows/action.yml
@@ -32,28 +32,8 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: clippy
-      - name: Run cargo clippy with no default features
-        uses: actions-rs/cargo@v1
-        with:
-          command: clippy
-          args: --no-default-features
-      - name: Run cargo clippy with aws enabled
-        uses: actions-rs/cargo@v1
-        with:
-          command: clippy
-          args: --features s3-storage
 
       - name: Run cargo test
         uses: actions-rs/cargo@v1
         with:
-          command: test
-      - name: Run cargo test with no default features
-        uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: --no-default-features
-      - name: Run cargo test with aws enabled
-        uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: --features s3-storage -- --test-threads=1
+          command: test -- --test-threads=1
