+++ htsget-http-actix/Cargo.toml
@@ -5,8 +5,7 @@ authors = ["Daniel del Castillo de la Rosa <delcastillodelarosadaniel@gmail.com>
 edition = "2021"
 
 [features]
-default = ["async", "htsget-search/default", "htsget-http-core/default"]
-async = ["tokio", "futures", "htsget-search/async", "htsget-http-core/async"]
+default = ["htsget-search/default", "htsget-http-core/default"]
 
 [dependencies]
 actix-web = "4.0"
@@ -17,8 +16,8 @@ futures-util = { version = "0.3", default-features = false }
 htsget-http-core = { path = "../htsget-http-core" }
 htsget-search = { path = "../htsget-search" }
 htsget-config = { path = "../htsget-config" }
-futures = { version = "0.3", optional = true }
-tokio = { version = "1.17", features = ["full"], optional = true }
+futures = { version = "0.3" }
+tokio = { version = "1.17", features = ["full"] }
 
 [dev-dependencies]
 htsget-test-utils = { path = "../htsget-test-utils" }
