+++ htsget-http-actix/src/handlers/get.rs
@@ -8,7 +8,7 @@ use actix_web::{
 use htsget_http_core::{get_response_for_get_request, Endpoint};
 use htsget_search::htsget::HtsGet;
 
-use crate::AsyncAppState;
+use crate::AppState;
 
 use super::handle_response;
 
@@ -16,7 +16,7 @@ use super::handle_response;
 pub async fn reads<H: HtsGet + Send + Sync + 'static>(
   request: Query<HashMap<String, String>>,
   path: Path<String>,
-  app_state: Data<AsyncAppState<H>>,
+  app_state: Data<AppState<H>>,
 ) -> impl Responder {
   let mut query_information = request.into_inner();
   query_information.insert("id".to_string(), path.into_inner());
@@ -34,7 +34,7 @@ pub async fn reads<H: HtsGet + Send + Sync + 'static>(
 pub async fn variants<H: HtsGet + Send + Sync + 'static>(
   request: Query<HashMap<String, String>>,
   path: Path<String>,
-  app_state: Data<AsyncAppState<H>>,
+  app_state: Data<AppState<H>>,
 ) -> impl Responder {
   let mut query_information = request.into_inner();
   query_information.insert("id".to_string(), path.into_inner());
