+++ htsget-http-actix/src/handlers/mod.rs
@@ -3,18 +3,13 @@ use actix_web::{http::StatusCode, Either, Responder};
 use htsget_http_core::{JsonResponse, Result};
 use pretty_json::PrettyJson;
 
-#[cfg(feature = "async")]
-pub use crate::handlers::async_handlers::{
+pub use crate::handlers::service_info::{
   get_service_info_json, reads_service_info, variants_service_info,
 };
 
-#[cfg(feature = "async")]
-pub mod async_handlers;
-pub mod blocking;
-#[cfg(feature = "async")]
 pub mod get;
-#[cfg(feature = "async")]
 pub mod post;
+pub mod service_info;
 
 mod pretty_json;
