+++ htsget-http-actix/src/main.rs
@@ -3,9 +3,6 @@ use std::env::args;
 use actix_web::{web, App, HttpServer};
 
 use htsget_config::config::{Config, USAGE};
-#[cfg(feature = "async")]
-use htsget_http_actix::async_configure_server as configure_server;
-#[cfg(not(feature = "async"))]
 use htsget_http_actix::configure_server;
 
 #[actix_web::main]
