+++ htsget-http-core/Cargo.toml
@@ -5,13 +5,12 @@ authors = ["Daniel del Castillo de la Rosa <delcastillodelarosadaniel@gmail.com>
 edition = "2021"
 
 [features]
-default = ["async", "htsget-search/default"]
-async = ["tokio", "futures", "htsget-search/async"]
+default = ["htsget-search/default"]
 
 [dependencies]
 thiserror = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 htsget-search = { path = "../htsget-search" }
 htsget-config = { path = "../htsget-config" }
-futures = { version = "0.3", optional = true }
-tokio = { version = "1.17", features = ["full"], optional = true }
+futures = { version = "0.3" }
+tokio = { version = "1.17", features = ["full"] }
